casting kcaltarget value to double to avoid error when value is read in as an int

+++ healthy_app/lib/screens/home/calorie_count.dart
@@ -81,8 +81,8 @@ class _CalorieCountState extends State<CalorieCount> {
   }
 
   double calculatePercentage() {
-    double kcalTarget = widget.calorieTarget;
-    double percentage = totalCalories.toDouble() / kcalTarget;
+    double kcalTarget = widget.calorieTarget.toDouble();
+    double percentage = (totalCalories.toDouble() / kcalTarget);
 
     print("Percentage " + (percentage * 100).toString() + "kcalTarget = " +
         kcalTarget.toString() + "calories = " + totalCalories.toString());
